@fcakyon 2021-12-24
Amazing fix @imyhxy! Do you have any intentions to open a pr with your yolov5 implementation?