@imyhxy 2021-12-24
@fcakyon Yes, after I finished the training experiments and if everything OK I will open a PR for it.