@ZwwWayne 2021-12-24
Hi @imyhxy ,
Thanks for your kind PR. Would you like to also check the case of numthread=1? Originally there are some OpenMMLab codebases using numthread=1. If 0 is the best, we can use that in MMDet.