@imyhxy 2021-12-24
@ZwwWayne Hi

I have done some test on another server which is the same as before but have 4 GPU.

|`setNumThreads`|Threads|%CPU us|%CPU sy|Times|
|:-:|:-:|:-:|:-:|:-:|
|0|188|33.1|4.3|0.213, 0.214, 0.213|
|1|188|33.4|4.3|0.215, 0.215, 0.216|
|2|220|36.8|5.4|0.215, 0.214, 0.213|
|4|284|39.3|7.0|0.214, 0.213, 0.215|
|8|412|44.5|11.4|0.223, 0.221, 0.223|
|16|668|50.8|23.3|0.232, 0.232, 0.232|

When setting `setNumThreads` to 0 or 1, there is no significant difference on **Thread Number**, **User CPU Time**, 
**System CPU Time** and **Training Time**, so they are pretty much is the same. But according to the document of opencv, the behaviour of `setNumThreads(1)` differs from framework, so use `setNumThreads(0)` is more stable and safety.

```
If threads == 0, OpenCV will disable threading optimizations and run all it's functions sequentially. Passing threads < 0 will reset threads number to system default. This function must be called outside of parallel region.

OpenCV will try to run its functions with specified threads number, but some behaviour differs from framework:

    * TBB - User-defined parallel constructions will run with the same threads number, if another is not specified. If later on user creates his own scheduler, OpenCV will use it.
    * OpenMP - No special defined behaviour.
    * Concurrency - If threads == 1, OpenCV will disable threading optimizations and run its functions sequentially.
    * GCD - Supports only values <= 0.
    * C= - No special defined behaviour.
```

The **Thread Number** and **CPU Time (us, sy)** increase as the **setNumThreads** increases, but the training speed is slowing down only when the total CPU workload exceed some point (~55% on my system). On the above experiments, the training speed gets some significant slow down when `setNumThreads()` to some number between 4-8 (I have not done such fine-grain setup).